last fix

+++ src/main/java/com/example/bandservice/service/BandService.java
@@ -6,9 +6,14 @@ import java.util.List;
 
 public interface BandService {
     Band create(Band band);
-    Band readById(Integer id);
+
+    Band readById(Long id);
+
     Band readByName(String name);
-    Band update(Integer id, Band band);
-    void delete(Integer id);
+
+    Band update(Long id, Band band);
+
+    void delete(Long id);
+
     List<Band> getAll();
 }
